all: retire unwanted aio brick for external build
AIO needs to be vastly re-implemented later in a different way, when
integrated with kernel upstream.

For now, use SIO instead of AIO when building an externel kernel
module for a kernel without the (obsolete) prepatch.

For now, AIO is nevertheless used when MARS_MAJOR is present,
which can happen only at a pre-patched kernel.